matrix_bot/requirements.txt
simplematrixbotlib==2.12.3
python-dotenv==0.21.1
httpx==0.28.1
matrix-nio[e2e]==0.24.0
